requirements.txt
eventlet==0.30.2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
Flask-Session==0.5.0
Flask-SocketIO==5.3.4
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==20.1.0
mysql-connector-python==8.0.33
PyJWT==2.7.0
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.7.2
pytz==2023.3
simple-websocket==0.10.1
websocket-client==1.5.3
Werkzeug==2.3.6
wheel== 0.40.0
